Initial commit after metadata reorganization into a single repository
The CABS tape digitization project was formerly divided into three
phases. These phases had more to do with project management than with
characteristics of the underlying dataset. Also, the original directory
layout made separation of data and metadata in such a way that all
metadata could be held in a single git repository difficult. (Version
control via git is practical for the metadata files but not for the
massive binary data files.) This is the first commit after a
reorganization to address these issues: Parameter files and calibration
files for all CABS tapes now reside in a single git repository.